+++ doc/CHANGES.rst
@@ -24,7 +24,7 @@ Features Added
 
 - Updated distributions:
 
-  - Products.ZCatalog = 2.13.1
+  - Products.ZCatalog = 2.13.2
 
 Restructuring
 +++++++++++++